+++ tests/testthat/setup.R
@@ -812,6 +812,9 @@ meth_tst <- list(
 # Suppress the warning for interaction terms being selected before all involved
 # main effects have been selected (only concerns L1 search):
 options(projpred.warn_L1_interactions = FALSE)
+# Suppress the warning thrown by proj_predict() in case of observation weights
+# that are not all equal to `1`:
+options(projpred.warn_wobs_ppd = FALSE)
 
 search_trms_tst <- list(
   default_search_trms = list(),